+++ src/DonorProfiles/resources/js/block/data/attributes.js
@@ -1,0 +1,12 @@
+/**
+ * Block Attributes
+*/
+
+const blockAttributes = {
+	enabled: {
+		type: 'array',
+		default: [],
+	},
+};
+
+export default blockAttributes;